backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiomysql==0.2.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 支付
wechatpayv3==1.2.35
alipay-sdk-python==3.7.166

# 腾讯云 COS
cos-python-sdk-v5==1.9.27

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.1

# 工具
python-dotenv==1.0.0
